refactor: extract backup storage helpers
